+++ install.py
@@ -9,7 +9,7 @@ import secrets
 
 from weecfg.extension import ExtensionInstaller
 
-VERSION = '0.12.1'
+VERSION = '0.12.2'
 
 
 def loader():